Use `all_engines` more consistently
This allows us remove some the nested loops from these tests.

Ideally `all_engines` would generate separate tests for each engine but
that can be a followup change.
